Updated run.sh, run_dev.sh and Makefile to make working environments

+++ scripts/deps-dn.sh
@@ -1,10 +1,12 @@
 #!/bin/bash
 set +x
 
-podman pod kill web
-podman pod kill mongo
-podman pod kill poussetaches
-podman pod rm web
-podman pod rm mongo 
-podman pod rm poussetaches 
-podman network rm microblogpub-network
+if podman network exists microblogpub-network; then
+    podman pod kill mongo
+    podman pod kill poussetaches
+    podman pod rm mongo 
+    podman pod rm poussetaches 
+    podman network rm microblogpub-network
+else
+    echo "microblogpub-network does not exists... exiting..."
+fi
